[SERV-1106] Add PTLS tool to pt-tools (#4)
- This PR includes much of the structure of how other command line tools
will be written for PT-tools
- PT-tools has both an api side as well as a cli side 
- the api is contained in pkg/pairtree
- the cli is called from main.go which accesses the cli in cmd/ptls
- Testing for both are in the same folders (as is normal in go
structure). Testing uses testifys assert as well as afero to create
temporary directories
- test-dir contains the test pairtree directory that most testing is
done on. This test-dir is copied in various tests
- pkg/errror-msgs contains error-msgs that are used through out the
program
- pkg/test-utls has various testing functoins that are used in different
test spread through different packages
- There will likely be things I change as we progress further in the
project. I'm debating how best to deal with the Logger , but for now,
having it in the ptls package works and is simple